[UPDATE] - create a file to store widgets of evaluation as build stars and heart
